colocando imutabilidade nos metodos do service Impl e fazendo a cobertura de testes para o service Impl

+++ src/main/java/com/pamela/desafiocompra/model/PaymentConditionRequest.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class PaymentConditionRequest {
 
     private Double inputValue;
-    private Double numberOfParcels;
+    private Integer numberOfParcels;
 }
